% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf.R
\name{Bf}
\alias{Bf}
\title{Bf (legacy function)}
\usage{
Bf(sd = 0.1, obtained = 0, dfdata = 99, uniform = 0, lower = 0,
  upper = 1, meanoftheory = 0, sdtheory = 1, tail = 2)
}
\arguments{
\item{sd}{observed standard error}

\item{obtained}{obtained mean}

\item{dfdata}{observed degrees of freedom}

\item{uniform}{is it uniform?}

\item{lower}{lower-bound (if uniform)}

\item{upper}{upper-bound (if uniform)}

\item{meanoftheory}{theory mean}

\item{sdtheory}{theory SD}

\item{tail}{number of tails}
}
\value{
list
}
\description{
The following R code is based on the Baguley and Kaye (2010) R code for the Dienes (2008) calculator, changing the likelihood function for the data from a Normal distribution to a t-distribution. This is a way of accounting for the variance of observations being unknown in advance of the data - thus no correction factor need be applied to the SE when this calculator is used.
From http://www.lifesci.sussex.ac.uk/home/Zoltan_Dienes/inference/Bf%20with%20t%20likelihood.html
}
