% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfrr.R
\name{bfrr}
\alias{bfrr}
\title{Bayes Factor with Robustness Region (normal)}
\usage{
bfrr(sample_mean = 0, sample_se = 0.1, sample_df = 99,
  model = "normal", ..., criterion = 3, rr_interval = NA,
  precision = 0.05)
}
\arguments{
\item{sample_mean}{the observed sample mean}

\item{sample_se}{the observed sample standard error}

\item{sample_df}{the observed sample degrees of freedom}

\item{model}{the model under which to calculate likelihood (H0, normal or uniform}

\item{...}{model parameters (mean, sd and tail for normal, lower and upper for uniform)}

\item{criterion}{the cutoff Bayes Factor for concluding evidence in favour of H0 or H1}

\item{rr_interval}{the parameter intervals within which to test for robustness}

\item{precision}{the step size for calculating the robustness region (default 0.05)}
}
\value{
list
}
\description{
Bayes Factor with Robustness Region (normal)
}
\examples{

rr <- bfrr(0.5, 1/sqrt(30), 29, "normal", mean = 0, sd = .7, tail = 1)
summary(rr)
plot(rr)
}
