% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default}
\alias{default}
\title{Give default value if doesn't exist}
\usage{
default(x, default, rm_na = TRUE)
}
\arguments{
\item{x}{The value to test if it exists}

\item{default}{The default value}

\item{rm_na}{Replace NA values with the default}
}
\value{
the value or default
}
\description{
Give default value if doesn't exist
}
\examples{
a <- 1
default(a, 2) # should return 1
remove(a)
default(a, 2) # should return 2

}
