% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Calculate likelihood of data under model}
\usage{
likelihood(sample_mean = 0, sample_se = 0.1, sample_df = 99,
  model = c("H0", "normal", "uniform"), ..., steps = 2000)
}
\arguments{
\item{sample_mean}{the observed sample mean}

\item{sample_se}{the observed sample standard error}

\item{sample_df}{the observed sample degrees of freedom}

\item{model}{the model under which to calculate likelihood (H0, normal or uniform}

\item{...}{model parameters (mean, sd and tail for normal, lower and upper for uniform)}

\item{steps}{the number of steps to integrate over}
}
\value{
double
}
\description{
Calculate likelihood of data under model
}
\examples{

lNull <- likelihood(0.5, 1/sqrt(30), 29, "H0")
lTheory <- likelihood(0.5, 1/sqrt(30), 29, "normal", mean = 0, sd = 1, tail = 1)
BayesFactor <- lNull/lTheory

}
