% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bfrr.R
\name{summary.bfrr}
\alias{summary.bfrr}
\title{Summarise bfrr object}
\usage{
\method{summary}{bfrr}(object, ...)
}
\arguments{
\item{object}{a bfrr list}

\item{...}{Other arguments}
}
\value{
string
}
\description{
Summarise bfrr object
}
